# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: string;
  image: string;
  category: string;
  subcategory: string;
  trending: boolean;
  bestSeller: boolean;
  nameEn: string;
  /** @nullable */
  badge?: string | null;
  tone: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: string;
  /** @minLength 1 */
  image: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  subcategory: string;
  trending: boolean;
  bestSeller: boolean;
  nameEn: string;
  /** @nullable */
  badge?: string | null;
  tone: string;
}

export interface Category {
  /** @minLength 1 */
  name: string;
  /**
     * @minItems 1
     * @items.minLength 1
     */
  subcategories: string[];
}

export interface SiteSettings {
  /** @minLength 1 */
  heroImage: string;
  /**
     * @minItems 2
     * @maxItems 2
     */
  heroHeadline: string[];
  heroDescription: string;
  /** @pattern ^[1-9][0-9]{7,14}$ */
  whatsappNumber: string;
}

export interface SiteSettingsInput {
  /** @minLength 1 */
  heroImage: string;
  /**
     * @minItems 2
     * @maxItems 2
     */
  heroHeadline: string[];
  heroDescription: string;
  /** @pattern ^[1-9][0-9]{7,14}$ */
  whatsappNumber: string;
}

export interface CategoriesInput {
  /** @minItems 1 */
  categories: Category[];
}

export interface CatalogResponse {
  products: Product[];
  categories: Category[];
  site: SiteSettings;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SessionResponse {
  authenticated: boolean;
}

export interface ErrorResponse {
  error: string;
}