# lib/api-zod/src/generated/types/catalogResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Product } from './product';
import type { SiteSettings } from './siteSettings';

export interface CatalogResponse {
  products: Product[];
  categories: Category[];
  site: SiteSettings;
}