# lib/api-zod/src/generated/types/categoriesInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface CategoriesInput {
  /** @minItems 1 */
  categories: Category[];
}