# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  /** @minLength 1 */
  name: string;
  /**
     * @minItems 1
     * @items.minLength 1
     */
  subcategories: string[];
}