# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: string;
  image: string;
  category: string;
  subcategory: string;
  trending: boolean;
  bestSeller: boolean;
  nameEn: string;
  /** @nullable */
  badge?: string | null;
  tone: string;
}