# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: string;
  /** @minLength 1 */
  image: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  subcategory: string;
  trending: boolean;
  bestSeller: boolean;
  nameEn: string;
  /** @nullable */
  badge?: string | null;
  tone: string;
}