# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  /** @minLength 1 */
  heroImage: string;
  /**
     * @minItems 2
     * @maxItems 2
     */
  heroHeadline: string[];
  heroDescription: string;
  /** @pattern ^[1-9][0-9]{7,14}$ */
  whatsappNumber: string;
}